Created core.h/core.cpp, added to makefiles. Started moving core structures from main to core beginning with COutPoint.
