+++ bitcoin-qt.pro
@@ -154,6 +154,7 @@ HEADERS += src/qt/bitcoingui.h \
     src/hash.h \
     src/uint256.h \
     src/serialize.h \
+    src/core.h \
     src/main.h \
     src/net.h \
     src/key.h \
@@ -234,6 +235,7 @@ SOURCES += src/qt/bitcoin.cpp \
     src/netbase.cpp \
     src/key.cpp \
     src/script.cpp \
+    src/core.cpp \
     src/main.cpp \
     src/init.cpp \
     src/net.cpp \
